# swagger.yml
swagger: "2.0"
info:
  description: "Especificação de API"
  version: "1.0.0"
  title: "RH Challenge API"
  contact:
    email: "lsaviofreitas@gmail.com"
host: "localhost:8080"
basePath: "/api"
tags:
- name: "Autenticação"
- name: "Setor"
  description: "Cadastro de Setores"
- name: "Pessoa"
  description: "Cadastro de Pessoas"
schemes:
- "http"
paths:
  /authenticate:
    post:
      tags:
      - "Autenticação"
      summary: "Realiza a autenticação do usuário"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: body
          name: "credentials"
          required: true
          schema:
            $ref: "#/definitions/AccountCredentials"
      responses:
        "200":
          description: "ok"
        "401":
          $ref: '#/components/responses/Unauthorized'
  /departaments:
    get:
      security:
        - bearerAuth: []
      tags:
      - "Setor"
      summary: "Retorna a lista de setores cadastrados"
      description: "Retorna a lista de setores cadastrados"
      operationId: "findAll"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Departament"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound' 
    post:
      security:
        - bearerAuth: []
      tags:
      - "Setor"
      summary: "Cadastra um novo setor"
      operationId: "create"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - in: "body"
          name: "departament"
          description: "Setor a ser cadastrado"
          required: true
          schema:
            $ref: "#/definitions/Departament"
      responses:
        "201":
          description: "created"
          schema:
            $ref: "#/definitions/Departament"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/BusinessError'
  /departaments/{id}:
    put:
      security:
        - bearerAuth: []
      tags:
      - "Setor"
      summary: "Atualizar setor"
      operationId: "update"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID do Setor a atualizar"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "departament"
          description: "Setor a ser atualizado"
          required: true
          schema:
            $ref: "#/definitions/Departament"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Departament"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/BusinessError'
    delete:
      security:
        - bearerAuth: []
      tags:
      - "Setor"
      summary: "Excluir setor"
      operationId: "delete"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID do Setor a excluir"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Departament"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/BusinessError'
    get:
      security:
        - bearerAuth: []
      tags:
      - "Setor"
      summary: "Recupera um Setor por id"
      operationId: "findById"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID do Setor a consultar"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Departament"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /people:
    get:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Retorna a lista de pessoas cadastradas"
      description: "Retorna a lista de pessoas cadastrados"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Person"
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound' 
    post:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Cadastra uma nova pessoa"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - in: "body"
          name: "departament"
          description: "Pessoa a ser cadastrada"
          required: true
          schema:
            $ref: "#/definitions/Person"
      responses:
        "201":
          description: "created"
          schema:
            $ref: "#/definitions/Departament"
        "401":
          $ref: '#/components/responses/Unauthorized'
  /people/{id}:
    put:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Atualizar setor"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID da Pessoa a atualizar"
          required: true
          type: "integer"
          format: "int64"
        - in: "body"
          name: "departament"
          description: "Setor a ser atualizado"
          required: true
          schema:
            $ref: "#/definitions/Person"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Person"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
    delete:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Excluir pessoa"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID do Setor a excluir"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Departament"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
    get:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Recupera uma Pessoa por id"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "id"
          in: "path"
          description: "ID da Pessoa a consultar"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Person"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /people/departament/{departamentId}:
    get:
      security:
        - bearerAuth: []
      tags:
      - "Pessoa"
      summary: "Consulta pessoas por setor"
      produces:
      - "application/json"
      parameters:
        - in: header
          name: Authorization
          type: string
          format: JWT
          required: true
        - name: "departamentId"
          in: "path"
          description: "ID do setor a consultar"
          required: true
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Person"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
definitions:
  Departament:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      description:
        type: "string"
    example:
      id: 1
      description: Secretaria
  Person:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      firstname:
        type: "string"
      lastname:
        type: "string"
      departament:
        $ref: "#/definitions/Departament"
  AccountCredentials:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
    example:
      username: ibyte
      password: password
components:
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Formato de Id inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BusinessError:
      description: Erro de negócio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message